MacroEconometrics

Documentation for MacroEconometrics.

MacroEconometrics.BayesianEstimated

MacroEconometrics.Estimated

MacroEconometrics.FixedEstimated

MacroEconometrics.VAR

MacroEconometrics._check_regularity_data

MacroEconometrics.is_stable

MacroEconometrics.is_stable

MacroEconometrics.make_companion_matrix

MacroEconometrics.make_companion_matrix

MacroEconometrics.BayesianEstimated — Type

Since conjugacy is incredibly rare and often not actually plausible (the priors are not plausible) in actual work, we will assume that all Bayesian estimated quantities are just samples, and can thus be put in an array. Additional meta data, such as warning information during sampling, etc can be put in the metadata field.

Behaviour

Most basic operations will directly be forwarded to the value field. As such, the type can be used like any other array, including multiplication with other arrays, iteration, etc.

Fields

value::Array{T}: The actual values of the estimated quantity. Should be an Array; If multiple chains have been used, then the chains should be stacked along the last dimension; That is, if only a single chain is being used, then the last dimension should be of length 1. So for a matrix B that is n×n, we would have an array of dimensions n×n×d×c where d are the number of draws, and c is the number of chains.

metadata::M: Any additional data that one wishes to save relating to the estimation. This could be warnings from the sampling algorithms, etc.

MacroEconometrics.Estimated — Type

The idea is to abstract he model from the way it has been estimated. In it's most essential form, a model is just a struct, with the same structure no matter how is has been estimated. The only time that estimation matter is when statistics about estimated quantities are asked for. Thus, any statistic that needs to be estimated in a model should be of type Estimated. Estimated itself can then be either Frequentist or Bayesian.

MacroEconometrics.FixedEstimated — Type

" Fixing the estimation.

Sometimes we want to fix a parameter to a specific value. Although this parameter is no-longer technically estimated, it is still an estimable quantity and thus is still a subtype of Estimated.

Fields

-value::Array{T}: The value of the quantity

MacroEconometrics.VAR — Type

VAR(n::Int, p::Int, B::E, b0::E, Σ::E, data::TSFrame) where {E<:Estimated}
VAR(data::TSFrame, p; type::Type{T}=BayesianEstimated)
VAR(n::Int, p::Int, B::FixedEstimated{T}, b0::FixedEstimated{T}, Σ::FixedEstimated{T}) where {T}

Every VAR model is of the form

\[y_t = b_0 + B_1 y_{t-1} + ... + B_p y_{t-p} + \varepsilon_t\]

where $B_i$ is $n \times n$ and $b_0$ is an $n \times 1$ vector. For the disturbance terms, $\varepsilon_t$ we generally assume $\varepsilon_t \sim N(0, \Sigma)$. This can therefore be equivelently written as

\[y_t = b_0 + B \tilde{y}_{t-1} + \varepsilon_t\]

where $B$ is now $n \times np$ and $\tilde{y}_{t-1}=(y_{t-1}', ..., y_{t-p}')'$.

General Information

VAR models are only approprite for regularly spaced observations. We therefore check whether this is the case here

Fields

n::Int: Number of variables in the VAR.

p::Int: Numbe of lags; See the equations above.

B::Union{Nothing, E<:Estimated}: Lag coefficient matrix; See the equations above; Should be a subtype of Estimated.

b0::Union{Nothing, E<:Estimated}: Intercept vector; See equations above.

Σ::Union{Nothing, E<:Estimated}: Covariance matrix of disturbance terms; See equation above.

data::Union{Nothing, TSFrame}: A time series data frame containing the data used for the VAR. Observations should be regularly spaced.

MacroEconometrics._check_regularity_data — Method

 Check whether data is of a regular and known frequency.

Arguments

data::TSFrame: Time series data

MacroEconometrics.is_stable — Method

is_stable(var::AbstractVectorAutoregression)

Check whether the VAR is stable.

Given a companionmatrix C of a VAR, the VAR is considered stable if and only if all eigenvalues of C are less than unity in absolute value.

Arguments

var::Estimated: The VAR model

References

Kilian, L., & Lütkepohl, H. (2017). Structural Vector Autoregressive Analysis: (1st ed.). Cambridge University Press. https://doi.org/10.1017/9781108164818

MacroEconometrics.is_stable — Method

is_stable(var::VAR{BayesianEstimated{T, M}}) where {T, M}

In case of Bayesian estimated VARs, the model is considered stable if and only if the VAR is stable for all parameter draws. As such, a rather strong view is taken here.

MacroEconometrics.make_companion_matrix — Method

make_companion(B::AbstractMatrix{T}) where {T}

Create the VAR companion matrix.

Given a VAR of the form

\[y_t = b_0 + B_1 y_{t-1} + ... + B_p y_{t-p} + \varepsilon_t\]

The companion matrix is given by

\[C = \begin{bmatrix} B_1 & B_2 & \dots & B_{p-1} & B_p \\ I_n & O & \dots & O & O \\ O & I_n & & O & O \\ \vdots & & \ddots & \vdots & \vdots \\ O & O & \dots & I_n & O \end{bmatrix}\]

Thus, $B$ is a $np\times np$ matrix.

Arguments

-B::AbstractMatrix{T}: Lag matrix in the form required for a VAR model. See the documentation of VAR.

References

Kilian, L., & Lütkepohl, H. (2017). Structural Vector Autoregressive Analysis: (1st ed.). Cambridge University Press. https://doi.org/10.1017/9781108164818

MacroEconometrics.make_companion_matrix — Method

make_companion_matrix(var::VAR{FixedEstimated{T}}) where {T}
make_companion_matrix(var::VAR{BayesianEstimated{T, M}}) where {T, M}

Make the companion matrix corresponding to a VAR model. In case of Bayesian estimated models, the companion matrix is constructed for each draw in each chain.