MacroEconometrics

Documentation for MacroEconometrics.

MacroEconometrics.BayesianEstimated

MacroEconometrics.Estimated

MacroEconometrics.FixedEstimated

MacroEconometrics.VAR

MacroEconometrics._check_regularity_data

MacroEconometrics.BayesianEstimated — Type

Since conjugacy is incredible rare and often not actually plausible (the priors are not plausible) in actual work, we will assume that all Bayesian estimated quantities are just samples, and can thus be put in an array. Additional meta data, such as warning information during sampling, etc can be put in the metadata field.

Behaviour

Most basic operations will directly be forwarded to the value field. As such, the type can be used like any other array, including multiplication with other arrays, iteration, etc.

Fields

value::Array{T}: The actual values of the estimated quantity. Should be an Array; If multiple chains have been used, then the chains should be stacked along the third dimension

metadata::M: Any additional data that one wishes to save relating to the estimation. This could be warnings from the sampling algorithms, etc.

MacroEconometrics.Estimated — Type

The idea is to abstract he model from the way it has been estimated. In it's most essential form, a model is just a struct, with the same structure no matter how is has been estimated. The only time that estimation matter is when statistics about estimated quantities are asked for. Thus, any statistic that needs to be estimated in a model should be of type Estimated. Estimated itself can then be either Frequentist or Bayesian.

MacroEconometrics.FixedEstimated — Type

" Fixing the estimation.

Sometimes we want to fix a parameter to a specific value. Although this parameter is no-longer technically estimated, it is still an estimable quantity and thus is still a subtype of Estimated.

Fields

-value::Array{T}: The value of the quantity

MacroEconometrics.VAR — Type

VAR(n::Int, p::Int, B::E, b0::E, Σ::E, data::TSFrame) where {E<:Estimated}
VAR(data::TSFrame, p; type::Type{T}=BayesianEstimated)
VAR(n::Int, p::Int, B::FixedEstimated{T}, b0::FixedEstimated{T}, Σ::FixedEstimated{T}) where {T}

Every VAR model is of the form

\[y_t = b_0 + B_1 y_{t-1} + ... + B_p y_{t-p} + \varepsilon_t\]

where $B_i$ is $n \times n$ and $b_0$ is an $n \times 1$ vector. For the disturbance terms, $\varepsilon_t$ we generally assume $\varepsilon_t \sim N(0, \Sigma)$. This can therefore be equivelently written as

\[y_t = b_0 + B \tilde{y}_{t-1} + \varepsilon_t\]

where $B$ is now $n \times np$ and $\tilde{y}_{t-1}=(y_{t-1}', ..., y_{t-p}')'$.

General Information

VAR models are only approprite for regularly spaced observations. We therefore check whether this is the case here

Fields

n::Int: Number of variables in the VAR.

p::Int: Numbe of lags; See the equations above.

B::Union{Nothing, E<:Estimated}: Lag coefficient matrix; See the equations above; Should be a subtype of Estimated.

b0::Union{Nothing, E<:Estimated}: Intercept vector; See equations above.

Σ::Union{Nothing, E<:Estimated}: Covariance matrix of disturbance terms; See equation above.

data::Union{Nothing, TSFrame}: A time series data frame containing the data used for the VAR. Observations should be regularly spaced.

MacroEconometrics._check_regularity_data — Method

 Check whether data is of a regular and known frequency.

Arguments

data::TSFrame: Time series data