MacroEconometrics

Documentation for MacroEconometrics.

MacroEconometrics.AbstractImpulseResponseFunction

MacroEconometrics.BayesianEstimated

MacroEconometrics.Estimated

MacroEconometrics.FixedEstimated

MacroEconometrics.FrequentistEstimated

MacroEconometrics.IRFCovarianceNormalisation

MacroEconometrics.IRFImpactNormalisation

MacroEconometrics.ImpulseResponseFunction

MacroEconometrics.IndependentNormalWishart

MacroEconometrics.OlsVAREstimator

MacroEconometrics.SVAR

MacroEconometrics.StructuralImpulseResponseFunction

MacroEconometrics.VAR

MacroEconometrics._check_regularity_data

MacroEconometrics._lag_ts

MacroEconometrics._simulate_VAR

MacroEconometrics._var_irf!

MacroEconometrics.create_minnesota_params

MacroEconometrics.estimate!

MacroEconometrics.is_stable

MacroEconometrics.is_stable

MacroEconometrics.make_companion_matrix

MacroEconometrics.make_companion_matrix

MacroEconometrics.simulate!

MacroEconometrics.simulate!

MacroEconometrics.stack_last_dim

MacroEconometrics.to_impact_normalisation

TSFrames.lag

MacroEconometrics.AbstractImpulseResponseFunction — Type

Should have a field called variables::Vector{Symbol} containing the variables

Should have a field called irfs::E<:Estimated containing the IRF estimates. Dimensions should be to×from×horizon

MacroEconometrics.BayesianEstimated — Type

Since conjugacy is incredibly rare and often not actually plausible (the priors are not plausible) in actual work, we will assume that all Bayesian estimated quantities are just samples, and can thus be put in an array. Additional meta data, such as warning information during sampling, etc can be put in the metadata field.

Behaviour

Most basic operations will directly be forwarded to the value field. As such, the type can be used like any other array, including multiplication with other arrays, iteration, etc.

Fields

value::Array{T}: The actual values of the estimated quantity. Should be an Array; If multiple chains have been used, then the chains should be stacked along the last dimension; That is, if only a single chain is being used, then the last dimension should be of length 1. So for a matrix B that is n×n, we would have an array of dimensions n×n×d×c where d are the number of draws, and c is the number of chains.

metadata::M: Any additional data that one wishes to save relating to the estimation. This could be warnings from the sampling algorithms, etc.

MacroEconometrics.Estimated — Type

The idea is to abstract he model from the way it has been estimated. In it's most essential form, a model is just a struct, with the same structure no matter how is has been estimated. The only time that estimation matter is when statistics about estimated quantities are asked for. Thus, any statistic that needs to be estimated in a model should be of type Estimated. Estimated itself can then be either Frequentist or Bayesian.

MacroEconometrics.FixedEstimated — Type

" Fixing the estimation.

Sometimes we want to fix a parameter to a specific value. Although this parameter is no-longer technically estimated, it is still an estimable quantity and thus is still a subtype of Estimated.

Fields

-value::Array{T}: The value of the quantity

MacroEconometrics.FrequentistEstimated — Type

Frequentist estimations usually consist of a point estimate and a confidence interval. If Bootstrapping is used, the confidence interval must not be symmetric. As such, having a separate lower and upper CI allows for asymmetric CIs.

Behaviour

Most operations are directly forwarded to the value field. For example, additiona, subtraction, ... are all forwarded to the value field. So is indexing.

Fields

value::Array{T}: The point estimates.

ci_lower::Array{T}: The lower end of the CI for each value in values

ci_upper::Array{T}: The upper end of the CI for each value in values

metadata::M: Any metadata. Should, for example, include the level of the CI.

MacroEconometrics.IRFCovarianceNormalisation — Type

Variance of structural shocks is normalised to unity.

MacroEconometrics.IRFImpactNormalisation — Type

Impact of structural shock on variable is normalised to unity. That is, each structural shock has a variable which it increases by one unit if the shock increases by one unit.

MacroEconometrics.ImpulseResponseFunction — Method

ImpulseResponseFunction(var::VAR{E}, horizon::Int) where {E <: BayesianEstimated}

Calculate the reduced form IRFs of a VAR model.

Arguments

var::VAR{<:BayesianEstimated}: A VAR model estimated using Bayesian estimation methods

horizon::Int: Maximum IRF horizon.

Retruns

Returns a tensor of dimension to×from×horizon×draw×chain. This can be flattened to a tensor of dimensions to×from×horizon×draw*chain using stack_last_dim.

MacroEconometrics.IndependentNormalWishart — Type

Estimate a VAR using a Independent Normal-Wishart prior.

Given a standard VAR, we can assume the following priors:

\[\begin{split} \beta \sim N(\beta_0, V_0) \\ \Sigma^{-1} \sim Wishart(S_0^{-1}, \nu_0) \end{split}\]

where $\beta$ is the vector of all VAR coefficients (each equations parameters stacked above each other) and where $\Sigma$ is the covariance matrix of the error terms.

This then results in the conditional posteriors

\[\begin{split} \beta | Y, \Sigma^{-1} &\sim N(\bar\beta, \bar{V}) \\ \bar{V} &= \left(V_0^{-1} + \sum_{t=1}^T Z_t'\Sigma^{-1}Z_t\right)^{-1} \\ \bar\beta &= \bar{V}\left(V_0^{-1}\beta_0 + \sum_{t=1}^T Z_t'\Sigma^{-1}y_t\right) \end{split}\]

and

\[\Sigma^{-1}|Y, \beta &\sim Wishart(\bar{S}^{-1}, \bar\nu) \\ \bar\nu &= T + \nu_0 \\ \bar{S} &= S_0 + \sum_{t=1}^T (y_t - Z_t\beta)(y_t - Z_t\beta)'\]

where $y_t$ is the outcome vector at time t, and where $Z_t = I_n \otimes (1, y_{t-1}', ..., y_{t-p}')$.

Fields

prior_β::AbsatractVector{T}: Prior mean of the coefficients.

prior_V::AbstractMatrix{T}: Prior covariance of coefficients.

prior_S::AbstractMatrix{T}: Prior scale matrix.

prior_ν::T: Prior degrees of freedom.

intercept::Bool: Estimate an intercept.

References

Koop, G., & Korobilis, D. (Eds.). (2010). Bayesian multivariate time series methods for emprirical macroeconomics. now.

MacroEconometrics.OlsVAREstimator — Type

OlsVAREstimator(intercept::Bool=true; confint_level=0.95)

OLS estimator for a VAR

Estimate VAR coefficients equation by equation using OLS. This uses the lm method in GLS.jl

Fields

intercept::Bool: Should an intercept be estimated?

confint_level::Real: Confidence level; Confidence intervals are symmetric and are using the asymptotic normal approximation.

References

Lütkepohl, H. (2005). New introduction to multiple time series analysis. New York: Springer. (page 72 below equation 3.2.12)

MacroEconometrics.SVAR — Type

SVAR(n::Int, p::Int, B::E, b0::E, Σ::E, data::TSFrame) where {E<:Estimated}
SVAR(data::TSFrame, p; type::Type{T}=BayesianEstimated)
SVAR(n::Int, p::Int, B::FixedEstimated{T}, b0::FixedEstimated{T}, Σ::FixedEstimated{T}) where {T}

Every SVAR model is of the form

\[A y_t = b_0 + B_1 y_{t-1} + ... + B_p y_{t-p} + \varepsilon_t\]

where $A$ is a $n \times n$ contemporanous matrix, $B_i$ is $n \times n$ and $b_0$ is an $n \times 1$ vector. For the disturbance terms, $\varepsilon_t$ we generally assume $\varepsilon_t \sim N(0, \Sigma)$. This can therefore be equivelently written as

\[A y_t = b_0 + B \tilde{y}_{t-1} + \varepsilon_t\]

where $B$ is now $n \times np$ and $\tilde{y}_{t-1}=(y_{t-1}', ..., y_{t-p}')'$.

General Information

SVAR models are only approprite for regularly spaced observations. We therefore check whether this is the case here

Fields

n::Int: Number of variables in the VAR.

p::Int: Numbe of lags; See the equations above.

A::Union{Nothing, E<:Estimated}: Contemporanous matrix; See the equations above;

B::Union{Nothing, E<:Estimated}: Lag coefficient matrix; See the equations above; Should be a subtype of Estimated.

b0::Union{Nothing, E<:Estimated}: Intercept vector; See equations above.

Σ::Union{Nothing, E<:Estimated}: Covariance matrix of disturbance terms; See equation above.

data::Union{Nothing, TSFrame}: A time series data frame containing the data used for the VAR. Observations should be regularly spaced.

MacroEconometrics.StructuralImpulseResponseFunction — Method

StructuralImpulseResponseFunction(var::VAR{<:BayesianEstimated}, horizon::Int, identification_method::CholeskyVAR)

Estimate Structural Impulse Response Functions using a Cholesky decomposition.

The reduced form covariance matrix is given by $A^{-1}\Sigma A^{-1}'$. If we normalise the structural covariance matrix $\Sigma = I$, then the reduced form covariance matrix is given by $A^{-1}A^{-1}'$. Thus, if we assume that $A$ takes a lower-triangular form, then we can estimate $A$ and $A^{-1}$ from the reduced form covariance matrix by obtaining the lower-triangular cholesky matrix of the reduced form covariance matrix. Since the covariance matrix is positive semi-definite, this lower-triangular matrix is unique.

Arguments

var::VAR{<:BayesianEstimated}: A VAR model estimated in a Bayesian way

horizon::Int: Maximum horizon of IRFs.

identification_method::CholeskyVAR: Identification method.

Returns

Returns a tensor of dimension to×from×horizon×draw×chain. This can be flattened to a tensor of dimensions to×from×horizon×draw*chain using stack_last_dim.

References

Kilian, L., & Lütkepohl, H. (2017). Structural Vector Autoregressive Analysis: (1st ed.). Cambridge University Press. https://doi.org/10.1017/9781108164818

MacroEconometrics.VAR — Type

VAR(n::Int, p::Int, B::E, b0::E, Σ::E, data::TSFrame) where {E<:Estimated}
VAR(data::TSFrame, p; type::Type{T}=BayesianEstimated)
VAR(n::Int, p::Int, B::FixedEstimated{T}, b0::FixedEstimated{T}, Σ::FixedEstimated{T}) where {T}

Every VAR model is of the form

\[y_t = b_0 + B_1 y_{t-1} + ... + B_p y_{t-p} + \varepsilon_t\]

where $B_i$ is $n \times n$ and $b_0$ is an $n \times 1$ vector. For the disturbance terms, $\varepsilon_t$ we generally assume $\varepsilon_t \sim N(0, \Sigma)$. This can therefore be equivelently written as

\[y_t = b_0 + B \tilde{y}_{t-1} + \varepsilon_t\]

where $B$ is now $n \times np$ and $\tilde{y}_{t-1}=(y_{t-1}', ..., y_{t-p}')'$.

General Information

VAR models are only approprite for regularly spaced observations. We therefore check whether this is the case here

Fields

n::Int: Number of variables in the VAR.

p::Int: Numbe of lags; See the equations above.

B::Union{Nothing, E<:Estimated}: Lag coefficient matrix; See the equations above; Should be a subtype of Estimated.

b0::Union{Nothing, E<:Estimated}: Intercept vector; See equations above.

Σ::Union{Nothing, E<:Estimated}: Covariance matrix of disturbance terms; See equation above.

data::Union{Nothing, TSFrame}: A time series data frame containing the data used for the VAR. Observations should be regularly spaced.

MacroEconometrics._check_regularity_data — Method

 Check whether data is of a regular and known frequency.

Arguments

data::TSFrame: Time series data

MacroEconometrics._lag_ts — Method

Create a lagged TSFrame.

Arguments

ts::TSFrame: TSFrame to lag

lag_value::Union{OrdinalRange{T, T}, AbstractVector{T}}: lags

Keyword Arguments

name_attach::String="__L": String to attach to variable names

remove_fist::Bool=true: Remove the first few rows tha will be missing for the lagged variables?

MacroEconometrics._simulate_VAR — Method

Simulate a Gaussian VAR

Arguments

B::AbstractMatrix{T}: Lag matrix of dimensions n×(n*p)

b0::AbstractVector{T}: Vector of intercepts of dimension n×1

Σ::AbstractMatrix{T}: Covariance matrix of errors of dimension n×n

initial::AbstractVector{T}: Initial values of dimension (n*p)×1; So it is the initial value of the companion form.

periods::Int: Number of periods to simulate for.

Keyword arguments

burnin::Int=100: Number of periods to use as burnin. This will reduce the dependency on the initial conditions.

rng::Random.AbstractRNG=Random.MersenneTwister(): Random number generator.

Returns

Returns a matrix Y of dimensions n×periods; Thus, rows are the variables, columns are the time points.

MacroEconometrics._var_irf! — Method

_var_irf!(irfs::AbstractArray, B::AbstractMatrix)

Calculate the IRF of a reduced form VAR model with lag coefficient matrix B. Save these results in irfs. irfs must have dimensions nvars×nvars×horizons

MacroEconometrics.create_minnesota_params — Method

create_minnesota_params(
    var::VAR{E},
    λ::T,
    θ::T;
    mean_first_own_lag::T=1.0,
    mean_other_lags::T=0.0,
    include_intercept::Bool=true,
    mean_intercept::T=0.0,
    variance_intercept::T=10.0
) where {E<:Estimated,T}

Crate the mean and covariance of a minnesota prior on the coefficients of a VAR. These can then be used as the prior mean anf variance of the VAR coefficients in the IndependentNormalWishart prior.

Arguments

var::VAR{<:Estimated}: a VAR model.

λ::<:Real: prior standard deviation on the first own lag.

θ::<:Real: relative tightness of the prior standard deviation in other lags. 0 < θ < 1

Keyword Arguments

mean_first_own_lag::<:Real=1.0: Mean on the first own lag.

mean_other_lags::<:Real=0.0: Mean on other lags than the first own lag.

include_intercept::Bool=true: Should an intercept be included. Intercepts are the first two elements in the returned mean vector.

mean_intercept::<:Real=0.0: Mean of the intercept.

variance_intercept::<:Real=10.0: Variance on the intercept. Commonly chosen quite large.

References

Kilian, L., & Lütkepohl, H. (2017). Structural Vector Autoregressive Analysis: (1st ed.). Cambridge University Press. https://doi.org/10.1017/9781108164818

MacroEconometrics.estimate! — Method

estimate!(var::AbstractVectorAutoregression, method::AbstractVAREstimator, args...; kwargs...)

Estimate a VAR using method.

Methods

Depending on the type of VAR, the methods can differ. Methods can also differ with regard to which statistical paradigm they are following. Some methods will estimate the VAR using Frequentist methods, while others will estimate the VAR using Bayesian methods.

Standard Gaussian VAR: VAR

This is represented by a standard VAR type. Methods currently include

OlsVAREstimator

IndependentNormalWishart

MacroEconometrics.is_stable — Method

is_stable(var::AbstractVectorAutoregression)

Check whether the VAR is stable.

Given a companionmatrix C of a VAR, the VAR is considered stable if and only if all eigenvalues of C are less than unity in absolute value.

Arguments

var::Estimated: The VAR model

References

Kilian, L., & Lütkepohl, H. (2017). Structural Vector Autoregressive Analysis: (1st ed.). Cambridge University Press. https://doi.org/10.1017/9781108164818

MacroEconometrics.is_stable — Method

is_stable(var::VAR{BayesianEstimated{T, M}}) where {T, M}

In case of Bayesian estimated VARs, the model is considered stable if and only if the VAR is stable for all parameter draws. As such, a rather strong view is taken here.

MacroEconometrics.make_companion_matrix — Method

make_companion(B::AbstractMatrix{T}) where {T}

Create the VAR companion matrix.

Given a VAR of the form

\[y_t = b_0 + B_1 y_{t-1} + ... + B_p y_{t-p} + \varepsilon_t\]

The companion matrix is given by

\[C = \begin{bmatrix} B_1 & B_2 & \dots & B_{p-1} & B_p \\ I_n & O & \dots & O & O \\ O & I_n & & O & O \\ \vdots & & \ddots & \vdots & \vdots \\ O & O & \dots & I_n & O \end{bmatrix}\]

Thus, $B$ is a $np\times np$ matrix.

Arguments

-B::AbstractMatrix{T}: Lag matrix in the form required for a VAR model. See the documentation of VAR.

References

Kilian, L., & Lütkepohl, H. (2017). Structural Vector Autoregressive Analysis: (1st ed.). Cambridge University Press. https://doi.org/10.1017/9781108164818

MacroEconometrics.make_companion_matrix — Method

make_companion_matrix(var::VAR{FixedEstimated{T}}) where {T}
make_companion_matrix(var::VAR{BayesianEstimated{T, M}}) where {T, M}

Make the companion matrix corresponding to a VAR model. In case of Bayesian estimated models, the companion matrix is constructed for each draw in each chain.

MacroEconometrics.simulate! — Method

Simulate a Gaussian SVAR.

Arguments

svar::SVAR{FixedEstimated{T}}: A SVAR model having FixedEstimated coefficients.

periods::Int: Number of periods to simulate the VAR for

Keyword Arguments

burnin::Int=100: Number of periods to use as burnin; The longer the burnin, the less the dependence on the initial values.

initial::AbstractVector{T}=zeros(T, svar.n*svar.p): Initial values. Should be in companion form, so a vector of length svar.n×svar.p

start_date::Date=Dates.today(): The simulated data will be stored in a TSFrame including dates. Thus, we need a start date.

frequency::DatePeriod=Dates.Quarter(1): Frequency of data.

rng::Random.AbstractRNG=Random.MersenneTwister(): Random Number Generator

Returns

Returns the same SVAR model. The data field in the SVAR model is overwritten with he new simulated data.

MacroEconometrics.simulate! — Method

Simulate a Gaussian VAR.

Arguments

var::VAR{FixedEstimated{T}}: A VAR model having FixedEstimated coefficients.

periods::Int: Number of periods to simulate the VAR for

Keyword Arguments

burnin::Int=100: Number of periods to use as burnin; The longe the burnin, the less the dependence on the initial values.

initial::AbstractVector{T}=zeros(T, var.n*var.p): Initial values. Should be in companion form, so a vector of length var.n×var.p

start_date::Date=Dates.today(): The simulated data will be stored in a TSFrame including dates. Thus, we need a start date.

frequency::DatePeriod=Dates.Quarter(1): Frequency of data.

rng::Random.AbstractRNG=Random.MersenneTwister(): Random Number Generator

Returns

Returns the same VAR model. The data field in the VAR model is overwritten witht he new simulated data.

MacroEconometrics.stack_last_dim — Method

stack_last_dim(arr::AbstractArray)

Stack the last dimension into the second to last dimension.

This is useful here, since we generally have the convention that the last dimension reflects the chain, and the second to last dimension the draws. As such, this function stacks the chains such that all draws of all chains are considered together.

Arguments

arr::AbstractArray

Returns

returns an array with one dimension less but with the second to last dimension enlarged by a factor equal the size of the last dimension of arr.

MacroEconometrics.to_impact_normalisation — Method

Move from covariance normalisation to impact normalisation.

Given a SVAR of the form

\[A y_t = b_0 + B_1 y_{t-1} + ... + B_p y_{t-p} + \Sigma^{1/2}\varepsilon\]

that has been covariance normalised, such that $\Sigma^{1/2}=I$, we can move to a SVAR that is impact normalised - the diagonal of $A$ consists of only ones, by premultiplying all matriced by $N = Diag(A)^{-1}$ where $Diag(X)$ denotes the matrix of the diagonal of $X$. The covariance matrix of the structural errors is then given by $NN'$.

Arguments

sirfs::<:StructuralImpulseResponseFunction obtained from a VAR or SVAR model and using any identification method relying on covariance normalisation.

TSFrames.lag — Method

If lag_value is a range or a vector of lags, then a vector of TSFrames will be returned where each element of the vector is one lagged TSFrame.