MacroEconometrics

Documentation for MacroEconometrics.

MacroEconometrics.AbstractImpulseResponseFunction

MacroEconometrics.BayesianEstimated

MacroEconometrics.Estimated

MacroEconometrics.FixedEstimated

MacroEconometrics.FrequentistEstimated

MacroEconometrics.IRFCovarianceNormalisation

MacroEconometrics.IRFImpactNormalisation

MacroEconometrics.OlsVAREstimator

MacroEconometrics.VAR

MacroEconometrics._check_regularity_data

MacroEconometrics._lag_ts

MacroEconometrics._simulate_VAR

MacroEconometrics.estimate!

MacroEconometrics.is_stable

MacroEconometrics.is_stable

MacroEconometrics.make_companion_matrix

MacroEconometrics.make_companion_matrix

MacroEconometrics.simulate!

TSFrames.lag

MacroEconometrics.AbstractImpulseResponseFunction — Type

Should have a field called variables::Vector{Symbol} containing the variables

Should have a field called irfs::E<:Estimated containing the IRF estimates. Dimensions should be to×from×horizon

MacroEconometrics.BayesianEstimated — Type

Since conjugacy is incredibly rare and often not actually plausible (the priors are not plausible) in actual work, we will assume that all Bayesian estimated quantities are just samples, and can thus be put in an array. Additional meta data, such as warning information during sampling, etc can be put in the metadata field.

Behaviour

Most basic operations will directly be forwarded to the value field. As such, the type can be used like any other array, including multiplication with other arrays, iteration, etc.

Fields

value::Array{T}: The actual values of the estimated quantity. Should be an Array; If multiple chains have been used, then the chains should be stacked along the last dimension; That is, if only a single chain is being used, then the last dimension should be of length 1. So for a matrix B that is n×n, we would have an array of dimensions n×n×d×c where d are the number of draws, and c is the number of chains.

metadata::M: Any additional data that one wishes to save relating to the estimation. This could be warnings from the sampling algorithms, etc.

MacroEconometrics.Estimated — Type

The idea is to abstract he model from the way it has been estimated. In it's most essential form, a model is just a struct, with the same structure no matter how is has been estimated. The only time that estimation matter is when statistics about estimated quantities are asked for. Thus, any statistic that needs to be estimated in a model should be of type Estimated. Estimated itself can then be either Frequentist or Bayesian.

MacroEconometrics.FixedEstimated — Type

" Fixing the estimation.

Sometimes we want to fix a parameter to a specific value. Although this parameter is no-longer technically estimated, it is still an estimable quantity and thus is still a subtype of Estimated.

Fields

-value::Array{T}: The value of the quantity

MacroEconometrics.FrequentistEstimated — Type

Frequentist estimations usually consist of a point estimate and a confidence interval. If Bootstrapping is used, the confidence interval must not be symmetric. As such, having a separate lower and upper CI allows for asymmetric CIs.

Behaviour

Most operations are directly forwarded to the value field. For example, additiona, subtraction, ... are all forwarded to the value field. So is indexing.

Fields

value::Array{T}: The point estimates.

ci_lower::Array{T}: The lower end of the CI for each value in values

ci_upper::Array{T}: The upper end of the CI for each value in values

metadata::M: Any metadata. Should, for example, include the level of the CI.

MacroEconometrics.IRFCovarianceNormalisation — Type

Variance of structural shocks is normalised to unity.

MacroEconometrics.IRFImpactNormalisation — Type

Impact of structural shock on variable is normalised to unity. That is, each structural shock has a variable which it increases by one unit if the shock increases by one unit.

MacroEconometrics.OlsVAREstimator — Type

OlsVAREstimator(intercept::Bool=true; confint_level=0.95)

OLS estimator for a VAR

Estimate VAR coefficients equation by equation using OLS. This uses the lm method in GLS.jl

Fields

intercept::Bool: Should an intercept be estimated?

confint_level::Real: Confidence level; Confidence intervals are symmetric and are using the asymptotic normal approximation.

References

Lütkepohl, H. (2005). New introduction to multiple time series analysis. New York: Springer. (page 72 below equation 3.2.12)

MacroEconometrics.VAR — Type

VAR(n::Int, p::Int, B::E, b0::E, Σ::E, data::TSFrame) where {E<:Estimated}
VAR(data::TSFrame, p; type::Type{T}=BayesianEstimated)
VAR(n::Int, p::Int, B::FixedEstimated{T}, b0::FixedEstimated{T}, Σ::FixedEstimated{T}) where {T}

Every VAR model is of the form

\[y_t = b_0 + B_1 y_{t-1} + ... + B_p y_{t-p} + \varepsilon_t\]

where $B_i$ is $n \times n$ and $b_0$ is an $n \times 1$ vector. For the disturbance terms, $\varepsilon_t$ we generally assume $\varepsilon_t \sim N(0, \Sigma)$. This can therefore be equivelently written as

\[y_t = b_0 + B \tilde{y}_{t-1} + \varepsilon_t\]

where $B$ is now $n \times np$ and $\tilde{y}_{t-1}=(y_{t-1}', ..., y_{t-p}')'$.

General Information

VAR models are only approprite for regularly spaced observations. We therefore check whether this is the case here

Fields

n::Int: Number of variables in the VAR.

p::Int: Numbe of lags; See the equations above.

B::Union{Nothing, E<:Estimated}: Lag coefficient matrix; See the equations above; Should be a subtype of Estimated.

b0::Union{Nothing, E<:Estimated}: Intercept vector; See equations above.

Σ::Union{Nothing, E<:Estimated}: Covariance matrix of disturbance terms; See equation above.

data::Union{Nothing, TSFrame}: A time series data frame containing the data used for the VAR. Observations should be regularly spaced.

MacroEconometrics._check_regularity_data — Method

 Check whether data is of a regular and known frequency.

Arguments

data::TSFrame: Time series data

MacroEconometrics._lag_ts — Method

Create a lagged TSFrame.

Arguments

ts::TSFrame: TSFrame to lag

lag_value::Union{OrdinalRange{T, T}, AbstractVector{T}}: lags

Keyword Arguments

name_attach::String="__L": String to attach to variable names

remove_fist::Bool=true: Remove the first few rows tha will be missing for the lagged variables?

MacroEconometrics._simulate_VAR — Method

Simulate a Gaussian VAR

Arguments

B::AbstractMatrix{T}: Lag matrix of dimensions n×(n*p)

b0::AbstractVector{T}: Vector of intercepts of dimension n×1

Σ::AbstractMatrix{T}: Covariance matrix of errors of dimension n×n

initial::AbstractVector{T}: Initial values of dimension (n*p)×1; So it is the initial value of the companion form.

periods::Int: Number of periods to simulate for.

Keyword arguments

burnin::Int=100: Number of periods to use as burnin. This will reduce the dependency on the initial conditions.

rng::Random.AbstractRNG=Random.MersenneTwister(): Random number generator.

Returns

Returns a matrix Y of dimensions n×periods; Thus, rows are the variables, columns are the time points.

MacroEconometrics.estimate! — Method

estimate!(var::AbstractVectorAutoregression, method::AbstractVAREstimator, args...; kwargs...)

Estimate a VAR using method.

Methods

Depending on the type of VAR, the methods can differ. Methods can also differ with regard to which statistical paradigm they are following. Some methods will estimate the VAR using Frequentist methods, while others will estimate the VAR using Bayesian methods.

Standard Gaussian VAR: VAR

This is represented by a standard VAR type. Methods currently include

OlsVAREstimator

MacroEconometrics.is_stable — Method

is_stable(var::AbstractVectorAutoregression)

Check whether the VAR is stable.

Given a companionmatrix C of a VAR, the VAR is considered stable if and only if all eigenvalues of C are less than unity in absolute value.

Arguments

var::Estimated: The VAR model

References

Kilian, L., & Lütkepohl, H. (2017). Structural Vector Autoregressive Analysis: (1st ed.). Cambridge University Press. https://doi.org/10.1017/9781108164818

MacroEconometrics.is_stable — Method

is_stable(var::VAR{BayesianEstimated{T, M}}) where {T, M}

In case of Bayesian estimated VARs, the model is considered stable if and only if the VAR is stable for all parameter draws. As such, a rather strong view is taken here.

MacroEconometrics.make_companion_matrix — Method

make_companion(B::AbstractMatrix{T}) where {T}

Create the VAR companion matrix.

Given a VAR of the form

\[y_t = b_0 + B_1 y_{t-1} + ... + B_p y_{t-p} + \varepsilon_t\]

The companion matrix is given by

\[C = \begin{bmatrix} B_1 & B_2 & \dots & B_{p-1} & B_p \\ I_n & O & \dots & O & O \\ O & I_n & & O & O \\ \vdots & & \ddots & \vdots & \vdots \\ O & O & \dots & I_n & O \end{bmatrix}\]

Thus, $B$ is a $np\times np$ matrix.

Arguments

-B::AbstractMatrix{T}: Lag matrix in the form required for a VAR model. See the documentation of VAR.

References

Kilian, L., & Lütkepohl, H. (2017). Structural Vector Autoregressive Analysis: (1st ed.). Cambridge University Press. https://doi.org/10.1017/9781108164818

MacroEconometrics.make_companion_matrix — Method

make_companion_matrix(var::VAR{FixedEstimated{T}}) where {T}
make_companion_matrix(var::VAR{BayesianEstimated{T, M}}) where {T, M}

Make the companion matrix corresponding to a VAR model. In case of Bayesian estimated models, the companion matrix is constructed for each draw in each chain.

MacroEconometrics.simulate! — Method

Simulate a Gaussian VAR.

Arguments

var::VAR{FixedEstimated{T}}: A VAR model having FixedEstimated coefficients.

periods::Int: Number of periods to simulate the VAR for

Keyword Arguments

burnin::Int=100: Number of periods to use as burnin; The longe the burnin, the less the dependence on the initial values.

initial::AbstractVector{T}=zeros(T, var.n*var.p): Initial values. Should be in companion form, so a vector of length var.n×var.p

start_date::Date=Dates.today(): The simulated data will be stored in a TSFrame including dates. Thus, we need a start date.

frequency::DatePeriod=Dates.Quarter(1): Frequency of data.

rng::Random.AbstractRNG=Random.MersenneTwister(): Random Number Generator

Returns

Returns the same VAR model. The data field in the VAR model is overwritten witht he new simulated data.

TSFrames.lag — Method

If lag_value is a range or a vector of lags, then a vector of TSFrames will be returned where each element of the vector is one lagged TSFrame.